Fork in test to work around classpath issues
Forking is needed to fix scalac classloader issues (see scala/scala-xml#20)
that conflate the compiler's run-time classpath and the compilation classpath.

The run-time (boot)classpath leaks into the compilation classpath,
so that scalac see classes that are used to run it
as classes used to compile against...

Forking uses a minimal classpath, so this craziness is avoided.

Alternatively, we could manage the scala instance as shown at the end of this file (commented)