Bump SBT and MiMa versions
Also note that we should bump sbt-osgi, which isn't quite
as simple as we need to deal with the AutoPlugin-ification
of that plugin.

After upgrading SBT, we can no longer append `Seq[File]` to
a `Seq[Credential]`. Not sure what changed, but I've been
more explicit rather than relying on the implicits.